Use Cynara helpers to obtain connection credentials
This makes the code simpler and makes it independent of non-upstream
GetConnectionSmackContext method which is going to be removed in the future.
Additionally, it is planned to implement credentials caching in the helper
library which will improve performance.
This commit also uses correct unique id to obtain credentials (sender's id
instead of service's id).

Change-Id: Ibacbd713a404418496649e4ea34e78dfa7c67cdb